+++ khoros/objects/__init__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+"""
+:Module:            khoros.objects.base
+:Synopsis:          This module contains sub-modules that are specific to the various Community API objects.
+:Usage:             ``from khoros.objects import base``
+:Example:           ``node_id = base.get_node_id('https://community.khoros.com/t5/Khoros-Blog/bg-p/relnote', 'blog')``
+:Created By:        Jeff Shurtliff
+:Last Modified:     Jeff Shurtliff
+:Modified Date:     22 Mar 2020
+"""
+
+__all__ = ['base']
+
+# Import the base module by default
+from . import base